backend/requirements.txt
google-adk>=0.0.1
google-cloud-aiplatform[adk,agent-engines]>=1.88.0
google-cloud-storage
websockets
fastapi 
uvicorn[standard] 
sqlalchemy 
pymysql
mysqlclient
psycopg2-binary
google-generativeai
asyncpg
qdrant_client
pandas
numpy
uv
sentence_transformers
fastapi-mcp
google-auth